+++ examples/test_import.py
@@ -1,2 +1,3 @@
 import amisim
+
 print(amisim.__version__)