+++ src/amisim/__init__.py
@@ -2,6 +2,7 @@
 Simple library functions for the project.
 """
 
+from .application import AmisimApplication
 from ._version import __version__
 
-__all__ = ["__version__"]
+__all__ = ["__version__", "AmisimApplication"]